autra_marketplace/requirements.txt
Django==5.0.1
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
stripe==7.8.0
Pillow==10.2.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
djangorestframework==3.14.0
gunicorn==21.2.0
django-cors-headers==4.3.1
boto3==1.34.23
django-storages==1.14.2
docker==7.0.0
